+++ src/main/java/com/tfc/smallerunits/utils/world/server/FakeServerLightingManager.java
@@ -26,7 +26,7 @@ public class FakeServerLightingManager extends ServerWorldLightManager {
 	
 	private ChunkManager manager;
 	private int lastSize = 0;
-	private int[] lighting = new int[0];
+	private byte[] lighting = new byte[0];
 	private int lastX = 0;
 	private int lastY = 0;
 	private int lastZ = 0;
@@ -46,7 +46,7 @@ public class FakeServerLightingManager extends ServerWorldLightManager {
 	public int tick(int toUpdateCount, boolean updateSkyLight, boolean updateBlockLight) {
 		FakeServerWorld world = (FakeServerWorld) manager.world;
 		if (lastSize != world.owner.unitsPerBlock) {
-			lighting = new int[world.owner.unitsPerBlock * world.owner.unitsPerBlock * world.owner.unitsPerBlock];
+			lighting = new byte[world.owner.unitsPerBlock * world.owner.unitsPerBlock * world.owner.unitsPerBlock];
 			lastSize = world.owner.unitsPerBlock;
 		}
 		int i;
@@ -115,7 +115,7 @@ public class FakeServerLightingManager extends ServerWorldLightManager {
 			}
 			max = Math.max(0, max);
 			if (lighting[toIndex(pos)] != (max)) {
-				lighting[toIndex(pos)] = (max);
+				lighting[toIndex(pos)] = (byte) max;
 				hasChanged = true;
 			}
 			return false;
